clarify submission with bots
Frontend: Clarify that the switch also applies to submissions. Longer
explanation text what happens when unchecked.
Backend: Do not send data to queue if bots were active (would be a
double-submit)
Documentation: Add documentation on the bots configuration parameter and
explain the behaviour of the Frontend's "Submit/Validate with Bots"
parameter